+++ CHANGELOG.md
@@ -77,6 +77,7 @@
 ### Performance Improvements
 
 - **shim:** Update kiennq-shim to v3.1.2 ([#6261](https://github.com/ScoopInstaller/Scoop/issues/6261))
+- **scoop-search:** Add JSON-based binary-index cache for 15-20x faster non-SQLite search ([#6657](https://github.com/ScoopInstaller/Scoop/pull/6657))
 
 ## [v0.5.2](https://github.com/ScoopInstaller/Scoop/compare/v0.5.1...v0.5.2) - 2024-07-26